feat: rpc0.2.1 definition migration comlete
